Hello,
Attached is a new version of new.ss.  Let me first point out two minor
changes in behavior:

First,
    (cond1 ((ten-or-twenty q)))
should only give 10; it used to give 20.  This is fixed.  Meanwhile,
    (cond1 (else (ten-or-twenty q)))
should still give 10 and 20, right?

Second, for my own mental consistency checking, I made "(cond*)", where
* stands for any of our cond varieties, mean "fail".

The big change is how I've tried to make "g", "(all g)", "(alli g)",
"(condi (g))", and "(cond@ (g))" all equivalent.  Think of the cond*
forms as if-statements, so everything but the very first goal should
be evaluation-delayed.  Moreover, we want (cond@ (g)) and (condi (g))
to be equivalent to g in all contexts, even in the position of a
function-call argument, such as (any always [ ]).  Hence cond* and all
should immediately evaluate the very first goal, without waiting for a
substitution to come in.  "==" is now a function.

Let me know what you think.  Is it easy for you to test the examples in
the book?  I'd recommend it...

Cheers,
	Ken